1 more chapter to go...
I was expecting this book to be just a "hey this is how you make things
in Java so here you go" but it was quite interesting and it challenged
me a few times. What started as me wanting to have something to go over
when I have time has resulted in me finding programming more fun and
rewarding.